refactor: split `command/reader` into submodules
