+++ src/debugger/command/reader/terminal.rs
@@ -1,48 +1,19 @@
-use std::fmt;
-use std::fs::{self, File};
-use std::io::{self, BufRead as _, BufReader, IsTerminal as _, Read as _, Write as _};
+use std::io::{BufRead as _, Write as _};
+use std::{fmt, fs, io};
+use std::{fs::File, io::BufReader};
 
 use crossterm::{cursor, execute, terminal};
 
-use crate::output::debugger_colors;
-use crate::term::{self, Key};
-use crate::{dprint, dprintln, output::Output};
-
-// TODO(refactor): Split file into multiple submodule files
-
-/// Read from argument first, if `Some`. Then read from stream.
-#[derive(Debug)]
-pub struct CommandReader {
-    argument: Option<Argument>,
-    stream: Stream,
-}
-
-/// Command-line argument.
-#[derive(Debug)]
-struct Argument {
-    buffer: String,
-    /// Byte index.
-    cursor: usize,
-}
-
-/// Stdin or interactive terminal.
-#[derive(Debug)]
-enum Stream {
-    Stdin(Stdin),
-    Terminal(Terminal),
-}
-
-/// Stdin which is not attached to a terminal, i.e. piped.
-#[derive(Debug)]
-struct Stdin {
-    stdin: io::Stdin,
-    /// Command must be stored somewhere to be referenced.
-    buffer: String,
-}
+use super::{Read, PROMPT};
+use crate::dprintln;
+use crate::{
+    output::{debugger_colors, Output},
+    term::{self, Key},
+};
 
 /// Interactive unbuffered terminal.
 #[derive(Debug)]
-struct Terminal {
+pub struct Terminal {
     stderr: io::Stderr,
     buffer: String,
     /// Byte index.
@@ -61,248 +32,6 @@ struct TerminalHistory {
     index: usize,
     /// `None` indicates failure to open file.
     file: Option<File>,
-}
-
-/// Must be ASCII to ensure `.len() == .chars().count()`
-const PROMPT: &str = "DEBUGGER> ";
-
-/// Print prompt and command.
-fn echo_command(command: Option<&str>) {
-    if command.is_some_and(|command| command.trim().is_empty()) {
-        return;
-    }
-    // Equivalent code found in `Terminal`
-    dprint!(Sometimes, Normal, "\x1b[1m{}", PROMPT);
-    dprintln!(
-        Sometimes,
-        Normal,
-        "{}",
-        command.unwrap_or("\x1b[3m(end of input)").trim()
-    );
-}
-
-/// A trait for objects which can yield a command, by iterating a string or reading a file.
-pub trait Read {
-    /// `None` indicates EOF.
-    /// Returned string slice MAY include leading or trailing whitespace.
-    fn read(&mut self) -> Option<&str>;
-}
-
-impl CommandReader {
-    pub fn from(argument: Option<String>) -> Self {
-        Self {
-            argument: argument.map(Argument::from),
-            stream: Stream::new(),
-        }
-    }
-}
-
-impl Read for CommandReader {
-    fn read(&mut self) -> Option<&str> {
-        // Always try to read from argument first
-        // If argument is `None`, or if read from argument returns `None`, then read from stream
-        // Note that `self.argument` cannot then be set to `None`, due to lifetime of returned value
-        if let Some(argument) = &mut self.argument {
-            if let Some(command) = argument.read() {
-                echo_command(Some(command));
-                return Some(command);
-            }
-        }
-        self.stream.read()
-    }
-}
-
-impl Argument {
-    pub fn from(source: String) -> Self {
-        Self {
-            buffer: source,
-            cursor: 0,
-        }
-    }
-}
-
-impl Read for Argument {
-    fn read(&mut self) -> Option<&str> {
-        // EOF
-        if self.cursor >= self.buffer.len() {
-            return None;
-        }
-
-        // Take characters until delimiter
-        let start = self.cursor;
-        let mut chars = self.buffer[self.cursor..].chars();
-        while let Some(ch) = chars.next().filter(|ch| *ch != '\n' && *ch != ';') {
-            self.cursor += ch.len_utf8();
-        }
-
-        let end = self.cursor;
-        self.cursor += 1; // sizeof('\n' or ';')
-
-        let command = self
-            .buffer
-            .get(start..end)
-            .expect("calculated incorrect character indexes");
-        Some(command)
-    }
-}
-
-impl Stream {
-    pub fn new() -> Self {
-        let stdin = io::stdin();
-        if stdin.is_terminal() {
-            Self::Terminal(Terminal::new())
-        } else {
-            Self::Stdin(Stdin::from(stdin))
-        }
-    }
-}
-
-impl Read for Stream {
-    fn read(&mut self) -> Option<&str> {
-        match self {
-            Self::Stdin(stdin) => {
-                let command = stdin.read();
-                echo_command(command);
-                command
-            }
-            Self::Terminal(terminal) => terminal.read(),
-        }
-    }
-}
-
-impl Stdin {
-    pub fn from(stdin: io::Stdin) -> Self {
-        Self {
-            stdin,
-            buffer: String::new(),
-        }
-    }
-
-    /// `None` indicates EOF.
-    fn read_char(&mut self) -> Option<char> {
-        read_char_from_bytes(|| self.read_byte()).expect("uh oh")
-    }
-
-    /// `None` indicates EOF.
-    fn read_byte(&mut self) -> Option<u8> {
-        let mut buf = [0; 1];
-        let bytes_read = self
-            .stdin
-            .read(&mut buf)
-            .expect("failed to read character from stdin");
-        if bytes_read == 0 {
-            return None;
-        }
-        Some(buf[0])
-    }
-}
-
-fn read_char_from_bytes<F>(mut next_byte: F) -> Result<Option<char>, ()>
-where
-    F: FnMut() -> Option<u8>,
-{
-    // TODO(refactor): Make this nicer
-    let Some(byte) = next_byte() else {
-        return Ok(None);
-    };
-
-    let mut bytes = [byte, 0, 0, 0];
-
-    let utf8_position = Utf8Position::from(byte);
-    let Some(utf8_len) = utf8_position.len() else {
-        return Err(());
-    };
-
-    for i in 1..utf8_len {
-        let Some(byte) = next_byte() else {
-            return Err(());
-        };
-        if !Utf8Position::from(byte).is_continuation() {
-            return Err(());
-        }
-        bytes[i] = byte;
-    }
-
-    let string = std::str::from_utf8(&bytes[0..utf8_len]).map_err(|_| ())?;
-    let mut chars = string.chars();
-    let ch = chars.next().ok_or(())?;
-    if chars.next().is_some() {
-        return Err(());
-    }
-    Ok(Some(ch))
-}
-
-/// Position of byte inside UTF-8 character
-enum Utf8Position {
-    /// First byte of 4-byte character
-    Begin4,
-    /// First byte of 3-byte character
-    Begin3,
-    /// First byte of 2-byte character
-    Begin2,
-    /// Only byte of 1-byte character
-    Begin1,
-    /// Continuation of multi-byte character
-    Continuation,
-}
-
-impl Utf8Position {
-    pub fn from(byte: u8) -> Self {
-        // TODO(refactor): Make this nicer
-        const MASK_4: u8 = 0b1111_0000;
-        const MASK_3: u8 = 0b1110_0000;
-        const MASK_2: u8 = 0b1100_0000;
-        const MASK_CONT: u8 = 0b1000_0000;
-        if byte & MASK_4 == MASK_4 {
-            return Utf8Position::Begin4;
-        }
-        if byte & MASK_3 == MASK_3 {
-            return Utf8Position::Begin3;
-        }
-        if byte & MASK_2 == MASK_2 {
-            return Utf8Position::Begin2;
-        }
-        if byte & MASK_CONT == MASK_CONT {
-            return Utf8Position::Continuation;
-        }
-        return Utf8Position::Begin1;
-    }
-
-    pub fn len(&self) -> Option<usize> {
-        match self {
-            Self::Begin4 => Some(4),
-            Self::Begin3 => Some(3),
-            Self::Begin2 => Some(2),
-            Self::Begin1 => Some(1),
-            Self::Continuation => None,
-        }
-    }
-
-    pub fn is_continuation(&self) -> bool {
-        matches!(self, Self::Continuation)
-    }
-}
-
-impl Read for Stdin {
-    fn read(&mut self) -> Option<&str> {
-        self.buffer.clear();
-
-        // Take characters until delimiter
-        loop {
-            let Some(ch) = self.read_char() else {
-                if self.buffer.is_empty() {
-                    return None; // First character is EOF
-                }
-                break;
-            };
-            if ch == '\n' || ch == ';' {
-                break;
-            }
-            self.buffer.push(ch);
-        }
-
-        Some(&self.buffer)
-    }
 }
 
 impl Terminal {
